Prototype_01/requirements.txt
# Core Framework
Flask

# Database ORM
Flask-SQLAlchemy

# Web Forms & Validation
Flask-WTF
email-validator  # Required by Flask-WTF for email fields

# AI Forecasting & Data Handling
pandas
prophet

# Configuration Management
python-dotenv  # For loading environment variables from .env file

# Email Notifications
Flask-Mail     # Helper for sending emails from Flask

# WSGI Server (for running the app in Docker/production)
gunicorn       # Common choice for Linux-based containers
